Move create work order modal to a new component
